0.6.0
many effect system improvements, upgrade UI implemented, many more particle effects

+++ Assets/Projectile.cs
@@ -8,7 +8,7 @@ public class Projectile : MonoBehaviour
     public Vector2 velocity;
     public Collider2D colider;
     public SpriteSet sprite;
-    public int despawnTime;
+    public float despawnTime;
     public float time;
     public bool canDespawn;
